pyproject.toml
[project]
name = "Hackathon-Phase2"
version = "0.1.0"
description = "Phase II - Task-todo"
requires-python = ">=3.11"
authors = [
    {name = "Team", email = "team@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "pydantic>=2.5.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
